backend/requirements.txt
aiohttp
asyncpg
python-dotenv
psycopg2-binary #librería para conectar Python con PostgreSQL
unidecode